Marvell SGMII patch: Fixed module output print
Marvell gave us an update to uboot for thunderbolt:
SDK_2.7.25_thunderbolt_v1.

Adding all their uboot changes would be unnecessary and high risk,
low reward. For this reason, we are only applying the minimal set
of changes to get SGMII working.

This patch comes from SDK_2.7.25_thunderbolt_v1 and is a required
precursor to be able to apply the one we really want:

"MHSLPONLGCY-151 Add SGMII support on GMAC1"

which is applied in go/fibercl/82370

Change-Id: Id6f62ada30c1a69a91c6cb43d9475457de902290
Signed-off-by: Eli Nidam <elini@marvell.com>